Skip 'best' Youtube formats when using built-in support.
The best, bestvideo and bestaudio formats can be added to
`preferred_fmt_ids` for use by the Youtube-DL extension with
manage_downloads. Hoever, the built-in support would fail if it
encountered one of these formats. Skipping allows "Download with
Youtube-DL" to use them in the same way it allows adaptive formats to be
used.