Do not try to rename the outtmpl file in the Youtube-DL extension.
The output file does not contain the escaped percents, resulting in a
file not found if renamed.

Youtube-DL also does not use '$' and escaping it is not required. An
episode title containing a '$' will produce a file containing two dollar
signs, and is then renamed to a single dollar sign. But a channel name
containing a dollar sign causes Youtube-DL to create a new channel
directory, with two dollar signs, for the output file. This directory
remains empty after gpodder moves the output file.

Both cases can be tested by renaming a channel to have '%$' and then
downloading episodes with Youtube-DL.

Therefore, dollar signs should not be escaped and renaming outtmpl is
not required as it will always match tempname, when assuming percents
are unescaped.