YoutubeDL extension should always rename partial files with extensions.
If a partial file exists with an extension appended to it, it should
always be renamed even if the partial file without an extension was
written to. Yt-dlp sometimes writes part of the file to the partial
file (bug), but always writes all of the file to a partial file with an
extension.