Prevent excessively wide confirmation dialogs.
The new undownload checkbox increased the width of the delete dialog.
Translations of the undownload checkbox, and translations of dialog
messages, could increase its width even more.

Wrapping the checkbox and message in scrolled windows allows their width
to be limited, similar to the show_message_details dialog. The textview
also visually separates the message from the checkbox, making it easier
to read.