Add/remove extension preferences buttons when loaded/unloaded.
The user must currently close preferences and reopen to configure
extensions that provide a preferences UI. This is a problem when the
user does not know the extension has preferences, but the sudden
appearance of the button when an extension is enabled might catch the
user's attention.

Furthermore, removing the button when the extension is unloaded prevents
any confusion as to why it is still there.