Add socks to settings (#1592)
Add a tab in Preferences called "Network"

The base config class has an observer facility. Using that we can
define an observer inside a closure containing a reference to the
UIConfig instance the application is using. And the observer will
update the global gpodder.config._proxies variable that various
network calls check to get the desired proxies.