Fix an empty `mygpo.device.uid` to avoid mygpod failures.
Some platforms could return an empty hostname string. This resulted in
mygpo always failing because it depends on the uid to be non-empty.

If hostname is empty, it now returns 'unknown', and appends a 1-8
digit identifier to the end of it to increase uniqueness of the uid.

Old installs are fixed by detecting an empty uid and generated a new
value.