extension tagging.py: make individual tags selectable
Make individual tags selectable - the subtitle field
(which gets written to subtitle and comments) is often
quite rediculously long.

Also change defaults.

New defaults:
    'strip_album_from_title': True,
    'genre_tag': 'Podcast',
    'always_remove_tags': False,
    'auto_embed_coverart': False,
    'set_artist_to_album': True,
    'set_version': 4,
    'modify_tags': True,
    'remove_before_modify': True,

    'write_title': True,
    'write_album': True,
    'write_subtitle': False,
    'write_comments': False,
    'write_genre': True,
    'write_pubdate': True,

While not as full featured as it could be, this will
be very compatible with more limited devices (such as rockbox players...)
and should be fairly sanitized.