bin/gpodder: Remove the redundant application uniqueness test
Gtk.Application ensures uniqueness without extra code.

Add a signal handler to 'notify::is-registered' which checks if the
current instance is remote (i.e. gPodder is already running). If the
instance is remote, logs the same message about Dbus activation as
before and handles the '--subscribe' command line option with the
running instance.