deviceplaylist.py: sanitize paths for os.path.ismount()
Weird things can happen if we don't change url-style
"file://some/path" to a real path, especially
if we don't find a mount point all the way down to
the root of the system. Could be an issue if you
point it towards a folder on your computer rather than
an actual external drive of some sort.

Strip both file: and mtp: prefixes from pathnames
for m3u files.