Device Sync: Use same filename as local by default
Add new config setting "use_title_as_filename", which
if set to true uses the episode title as the device filename.

If false (default), use the same filename on device as local.
This should help prevent issues with text encoding in filenames,
for instance in rockbox players.

For the playlist, get the actual title of the episode for the
extinf line rather than using the title from the filename.

Also shorten max filename length considerably. Set it to 60,
was 120.