Change progress dialog to be non-modal.
The main window is unresponsive during long running actions on the main
loop. The progress dialog appeared to notify the user of the action's
progress.

However, if the action failed and opened another dialog, it would cause
the progress dialog to appear due to exceeding the timeout. This caused
a deadlock between two modal dialogs, where the error dialog couldn't be
clicked and the progress dialog had no cancel button, or was maybe also
not clickable.

Making the progress dialog non-modal should prevent this from happening.
The user will be able to click the non-responsive UI behind the dialog,
but that is better than gpodder locking up.