Use DBus ShowItems method in on_open_episode_download_folder()
Use the org.freedesktop.FileManager1.ShowItems method to open the
selected episode in the default file manager. The file manager then
opens with the given episode already selected.

If a Dbus connection cannot be obtained or the method call fails, fall
back to opening the containing folder with util.gui_open(), as before.

Add debug logging to indicate which method is used.