add observers to Resolver; add AutoRegisterObserver
observers are notified when an implementation is added or removed
from a Resolver.

AutoRegisterObserver is an utility class to handle (un)registration
of listeners when an ObservableService implementation is added or
removed from a Resolver in gPodder's registry.

This is useful to always be notified of eg. a "PlayerInterface"
ObservableService when it emits PlaybackStarted/PlaybackStopped/...
signals. It might be implemented in an extension: appearing at any time
(the user enabled the extension) and disapearing and being replaced by
another extension.